% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_fun.R
\name{permute_fun}
\alias{permute_fun}
\title{This function defines the permutation strategry}
\usage{
permute_fun(data, mle)
}
\arguments{
\item{data}{A data frame specifying the data to be analysed.}

\item{mle}{A string that indicated response variable}
}
\value{
A data frame containing the data with a permuted y.
}
\description{
perform permuation on response variable i.e. y,
using the stop criterion suggested by Anscombe
}
\examples{
data<-mtcars
permute<-permute_fun(data=data,mle='mpg')
}
