% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1.R
\name{s1}
\alias{s1}
\title{Return t values of general linear model}
\usage{
s1(data, formula)
}
\arguments{
\item{data}{A dataframe specifying the data to be analysed}

\item{formula}{A formula in the form like'y~x1+x2+x3 in lm function}
}
\value{
An object of class "lm"
}
\description{
perform mixed linear regression in lmer/lmertest package
for getting observed t values or permutation test
}
\examples{
data<-mtcars
formula<-mpg~cyl
s1(data=data,formula=formula)
}
