% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{s2}
\alias{s2}
\title{Estimate t values of mixed linear model}
\usage{
s2(data, formula)
}
\arguments{
\item{data}{A dataframe specifying the data to be analysed}

\item{formula}{A formula in the form like'y~x1+x2+x3+(1|x4) or like'y~x1+x2+x3
+(x3|x4) in lmer function}
}
\value{
An object of class "lmerTest"
}
\description{
perform mixed linear regression in lmerTest package for getting
observed t values or permutation test
}
\examples{
data<-mtcars
formula<-mpg~cyl+(1|gear)
statistic<-s2(data=data,formula=formula)
}
